New Changes

+++ main/torrentvpn.sh
@@ -3,10 +3,7 @@ nmcli con show --active | grep -Ei 'wireguard|vpn|tun'
 vpnCode="${PIPESTATUS[1]}"
 pgrep -i 'torrent'
 torrentCode=$?
-if [[ $vpnCode == 1 ]] && [[ $torrentCode == 0 ]]; then
-  pkill -i 'torrent'
-  echo 'torrent killed'
-else
-  echo ''
-fi
+[[ $vpnCode == 1  && $torrentCode == 0 ]] && pkill -i 'torrent' && echo 'torrent killed'
+  
+
